+++ FreeRTOS-Plus/Demo/AWS/Fleet_Provisioning_Windows_Simulator/Fleet_Provisioning_With_CSR_Demo/DemoTasks/FleetProvisioningDemoExample.c
@@ -80,7 +80,7 @@
 /* Demo includes. */
 #include "mqtt_operations.h"
 #include "pkcs11_operations.h"
-#include "fleet_provisioning_tinycbor_serializer.h"
+#include "tinycbor_serializer.h"
 #include "using_mbedtls_pkcs11.h"
 
 /**
