+++ FreeRTOS-Plus/Demo/AWS/Fleet_Provisioning_Windows_Simulator/Fleet_Provisioning_With_CSR_Demo/tinycbor_serializer.c
@@ -37,7 +37,7 @@
 #include "fleet_provisioning.h"
 
 /* Header include. */
-#include "fleet_provisioning_tinycbor_serializer.h"
+#include "tinycbor_serializer.h"
 /*-----------------------------------------------------------*/
 
 bool xGenerateCsrRequest( uint8_t * pucBuffer,
